hotfix

+++ api/modules/edit_product/edit_product.php
@@ -2,22 +2,25 @@
 
 $product_id = $_POST['product_id'];
 
-$downloadedImages = $_POST['downloadedImages'];
 
-$imagesInDB =  array_column(
-    $mysqli->query("SELECT id FROM products_media WHERE product_id = $product_id")->fetch_all(MYSQLI_ASSOC),
-    'id'
-);
+if (isset($_POST['downloadedImages'])) {
+    $downloadedImages = $_POST['downloadedImages'];
 
-$diff = array_diff(
-    $imagesInDB,
-    $downloadedImages,
-);
+    $imagesInDB =  array_column(
+        $mysqli->query("SELECT id FROM products_media WHERE product_id = $product_id")->fetch_all(MYSQLI_ASSOC),
+        'id'
+    );
 
-foreach ($diff as $key => $image_id) {
-    $image_name = $mysqli->query("SELECT name FROM products_media WHERE id = $image_id")->fetch_assoc();
-    $deleteRes = $mysqli->query("DELETE FROM products_media WHERE id = $image_id");
-    $path_to_file = $config['uploaddir'] . "/" . $image_name['name'];
+    $diff = array_diff(
+        $imagesInDB,
+        $downloadedImages,
+    );
+
+    foreach ($diff as $key => $image_id) {
+        $image_name = $mysqli->query("SELECT name FROM products_media WHERE id = $image_id")->fetch_assoc();
+        $deleteRes = $mysqli->query("DELETE FROM products_media WHERE id = $image_id");
+        $path_to_file = $config['uploaddir'] . "/" . $image_name['name'];
+    }
 }
 
 foreach ($_FILES as $not_named_variable_name => $file) {
